\name{as_spatial_dataframes}
\alias{as_spatial_dataframes}
\title{Extract spatialpoint and spatialline dataframes for vertices and edges}
\usage{
as_spatial_dataframes(np)
}
\arguments{
  \item{np}{a NetworkPlan}
}
\value{
a list with vertices spatialdataframe and edges
spatiallinesdataframe
}
\description{
Extract spatialpoint and spatialline dataframes for
vertices and edges
}

