\name{default_selector}
\alias{default_selector}
\title{Default selector for sequence method
This simply selects the node with the smallest vertex id}
\usage{
default_selector(node_df)
}
\arguments{
  \item{node_df}{SpatialPointsDataFrame to apply selection
  to}
}
\value{
subset of the node_df
}
\description{
Default selector for sequence method This simply selects
the node with the smallest vertex id
}

