\docType{data}
\name{default_sequence_model}
\alias{default_sequence_model}
\title{Default rollout_functions for sequence_plan_far}
\format{\preformatted{List of 2
 $ accumulator:function (node_df, edge_df, g, vid)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 160 24 162 1 24 1 160 162
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x6d6cc80> 
 $ selector   :function (node_df)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 148 21 150 1 21 1 148 150
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x6d6cc80> 
}}
\usage{
default_sequence_model
}
\description{
Default rollout_functions for sequence_plan_far
}
\keyword{datasets}

