\name{igraph_to_json}
\alias{igraph_to_json}
\title{Convert an igraph to json of form { vertices: {...}, edges: {...} }}
\usage{
igraph_to_json(ig)
}
\arguments{
  \item{ig}{igraph}
}
\value{
json representation of an igraph
}
\description{
Convert an igraph to json of form { vertices: {...}, edges:
{...} }
}

