\name{is_valid_networkplan}
\alias{is_valid_networkplan}
\title{Determine whether a NetworkPlan is valid}
\usage{
is_valid_networkplan(np)
}
\arguments{
  \item{np}{a NetworkPlan}
}
\value{
A boolean indicating whether the NetworkPlan is valid
}
\description{
Determine whether a NetworkPlan is valid
}

