\name{json_to_igraph}
\alias{json_to_igraph}
\title{Convert json of form { vertices: {...}, edges: {...} }
to an igraph}
\usage{
json_to_igraph(json_str)
}
\arguments{
  \item{json_str}{json representation of graph as a string}
}
\value{
igraph
}
\description{
Convert json of form { vertices: {...}, edges: {...} } to
an igraph
}

