% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pop_accumulate}
\alias{pop_accumulate}
\title{"module" maintaining various downstream accumulator and
sequence selector functions for use with the NetworkPlanner.R
\code{sequence_plan_far} function}
\usage{
pop_accumulate(node_df, edge_df, g, vid)
}
\description{
"module" maintaining various downstream accumulator and
sequence selector functions for use with the NetworkPlanner.R
\code{sequence_plan_far} function
}
\seealso{
\code{\link{sequence_plan_far}}
TODO:  incorporate validation check to determine whether model
      will work with plan (i.e. does it have the right attributes)
}

