% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read_networkplan}
\alias{read_networkplan}
\title{Read scenario from a directory in the filesystem and
create it's NetworkPlan.  Converts scenario into an igraph object}
\usage{
read_networkplan(directory_name)
}
\arguments{
\item{directory_name}{absolute or relative path to a directory from which a NetworkPlan is loaded}
}
\value{
A NetworkPlan object
}
\description{
Read scenario from a directory in the filesystem and
create it's NetworkPlan.  Converts scenario into an igraph object
}

