% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{default_sequence_model}
\alias{default_sequence_model}
\title{Default rollout_functions for sequence_plan_far}
\format{\preformatted{List of 2
 $ accumulator:function (node_df, edge_df, g, vid)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 255 24 257 1 24 1 255 257
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0xd4cb760> 
 $ selector   :function (node_df)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 243 21 245 1 21 1 243 245
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0xd4cb760> 
}}
\usage{
default_sequence_model
}
\description{
Default rollout_functions for sequence_plan_far
}
\keyword{datasets}

