\docType{data}
\name{default_sequence_model}
\alias{default_sequence_model}
\title{Default rollout_functions for sequence_plan_far}
\format{\preformatted{List of 2
 $ accumulator:function (node_df, edge_df, g, vid)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 278 24 280 1 24 1 278 280
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x61adb68> 
 $ selector   :function (node_df)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 266 21 268 1 21 1 266 268
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x61adb68> 
}}
\usage{
default_sequence_model
}
\description{
Default rollout_functions for sequence_plan_far
}
\keyword{datasets}

