\docType{data}
\name{default_sequence_model}
\alias{default_sequence_model}
\title{Default rollout_functions for sequence_plan_far}
\format{\preformatted{List of 2
 $ accumulator:function (node_df, edge_df, g, vid)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 201 24 203 1 24 1 201 203
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0xa673518> 
 $ selector   :function (node_df)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 189 21 191 1 21 1 189 191
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0xa673518> 
}}
\usage{
default_sequence_model
}
\description{
Default rollout_functions for sequence_plan_far
}
\keyword{datasets}

