\name{download_scenario}
\alias{download_scenario}
\title{Download scenario from networkplanner into the given directory}
\usage{
download_scenario(scenario_number, directory_name = NULL)
}
\arguments{
  \item{scenario_number}{scenario_number on
  http://networkplanner.modilabs.org}

  \item{directory_name}{path to write the downloaded
  scenario into. By default, and if directory_name==NULL,
  directory_name will be scenario number within current
  working directory}
}
\description{
Download scenario from networkplanner into the given
directory
}

