\name{read_networkplan}
\alias{read_networkplan}
\title{Read network plan from a directory in the filesystem}
\usage{
read_networkplan(directory_name, debug = F)
}
\arguments{
  \item{directory_name}{absolute or relative path to a
  directory from which a NetworkPlan is loaded}

  \item{debug}{if TRUE, will verify inputs and run
  failsafes}
}
\value{
A NetworkPlan object
}
\description{
Read network plan from a directory in the filesystem
}

